requirements.txt
mariadb>=1.1.10
openai>=1.0.0
tiktoken>=0.7.0
mcp>=1.0.0
fastmcp>=2.0.0
streamlit>=1.30.0
llama-index-core>=0.10.0
llama-index-readers-file>=0.1.0
pytest>=8.0.0
